+++ example/src/main/java/com/h6ah4i/android/example/advrecyclerview/demo_e_already_expanded/AlreadyExpandedGroupsExpandableExampleAdapter.java
@@ -182,6 +182,14 @@ class AlreadyExpandedGroupsExpandableExampleAdapter
         return true;
     }
 
+    @Override
+    public boolean getInitialGroupExpandedState(int groupPosition) {
+        // NOTE:
+        // This method can also be used to control initial state of group items.
+        // Make sure to call `setDefaultGroupsExpandedState(false)` to take effect.
+        return false;
+    }
+
     // NOTE: This method is called from Fragment
     public void addGroupAndChildItemsBottom(int groupCount, int childCount) {
         int groupPosition = mProvider.getGroupCount();