+++ library/src/main/java/com/h6ah4i/android/widget/advrecyclerview/expandable/ExpandableRecyclerViewWrapperAdapter.java
@@ -65,7 +65,7 @@ class ExpandableRecyclerViewWrapperAdapter
 
         mExpandableItemAdapter = getExpandableItemAdapter(adapter);
         if (mExpandableItemAdapter == null) {
-            throw new IllegalArgumentException("adapter does not implement RecyclerViewExpandableListManager");
+            throw new IllegalArgumentException("adapter does not implement ExpandableItemAdapter");
         }
 
         if (manager == null) {
@@ -75,7 +75,10 @@ class ExpandableRecyclerViewWrapperAdapter
         mExpandableListManager = manager;
 
         mPositionTranslator = new ExpandablePositionTranslator();
-        mPositionTranslator.build(mExpandableItemAdapter, mExpandableListManager.getDefaultGroupsExpandedState());
+        mPositionTranslator.build(
+                mExpandableItemAdapter,
+                ExpandablePositionTranslator.BUILD_OPTION_DEFAULT,
+                mExpandableListManager.getDefaultGroupsExpandedState());
 
         if (expandedItemsSavedState != null) {
             // NOTE: do not call hook routines and listener methods
@@ -206,7 +209,10 @@ class ExpandableRecyclerViewWrapperAdapter
     private void rebuildPositionTranslator() {
         if (mPositionTranslator != null) {
             long[] savedState = mPositionTranslator.getSavedStateArray();
-            mPositionTranslator.build(mExpandableItemAdapter, mExpandableListManager.getDefaultGroupsExpandedState());
+            mPositionTranslator.build(
+                    mExpandableItemAdapter,
+                    ExpandablePositionTranslator.BUILD_OPTION_DEFAULT,
+                    mExpandableListManager.getDefaultGroupsExpandedState());
 
             // NOTE: do not call hook routines and listener methods
             mPositionTranslator.restoreExpandedGroupItems(savedState, null, null, null);
@@ -691,9 +697,9 @@ class ExpandableRecyclerViewWrapperAdapter
         }
 
         if (expand) {
-            expandGroup(groupPosition, true);
+            expandGroup(groupPosition, true, null);
         } else {
-            collapseGroup(groupPosition, true);
+            collapseGroup(groupPosition, true, null);
         }
 
         return true;
@@ -701,25 +707,31 @@ class ExpandableRecyclerViewWrapperAdapter
 
     /*package*/ void expandAll() {
         if (!mPositionTranslator.isEmpty() && !mPositionTranslator.isAllExpanded()) {
-            mPositionTranslator.build(mExpandableItemAdapter, true);
+            mPositionTranslator.build(
+                    mExpandableItemAdapter,
+                    ExpandablePositionTranslator.BUILD_OPTION_EXPANDED_ALL,
+                    mExpandableListManager.getDefaultGroupsExpandedState());
             notifyDataSetChanged();
         }
     }
 
     /*package*/ void collapseAll() {
         if (!mPositionTranslator.isEmpty() && !mPositionTranslator.isAllCollapsed()) {
-            mPositionTranslator.build(mExpandableItemAdapter, false);
+            mPositionTranslator.build(
+                    mExpandableItemAdapter,
+                    ExpandablePositionTranslator.BUILD_OPTION_COLLAPSED_ALL,
+                    mExpandableListManager.getDefaultGroupsExpandedState());
             notifyDataSetChanged();
         }
     }
 
-    /*package*/ boolean collapseGroup(int groupPosition, boolean fromUser) {
+    /*package*/ boolean collapseGroup(int groupPosition, boolean fromUser, Object payload) {
         if (!mPositionTranslator.isGroupExpanded(groupPosition)) {
             return false;
         }
 
         // call hook method
-        if (!mExpandableItemAdapter.onHookGroupCollapse(groupPosition, fromUser)) {
+        if (!mExpandableItemAdapter.onHookGroupCollapse(groupPosition, fromUser, payload)) {
             return false;
         }
 
@@ -736,24 +748,24 @@ class ExpandableRecyclerViewWrapperAdapter
             final long packedPosition = ExpandableAdapterHelper.getPackedPositionForGroup(groupPosition);
             final int flatPosition = mPositionTranslator.getFlatPosition(packedPosition);
 
-            notifyItemChanged(flatPosition);
+            notifyItemChanged(flatPosition, payload);
         }
 
         // raise onGroupCollapse() event
         if (mOnGroupCollapseListener != null) {
-            mOnGroupCollapseListener.onGroupCollapse(groupPosition, fromUser);
+            mOnGroupCollapseListener.onGroupCollapse(groupPosition, fromUser, payload);
         }
 
         return true;
     }
 
-    /*package*/ boolean expandGroup(int groupPosition, boolean fromUser) {
+    /*package*/ boolean expandGroup(int groupPosition, boolean fromUser, Object payload) {
         if (mPositionTranslator.isGroupExpanded(groupPosition)) {
             return false;
         }
 
         // call hook method
-        if (!mExpandableItemAdapter.onHookGroupExpand(groupPosition, fromUser)) {
+        if (!mExpandableItemAdapter.onHookGroupExpand(groupPosition, fromUser, payload)) {
             return false;
         }
 
@@ -769,12 +781,12 @@ class ExpandableRecyclerViewWrapperAdapter
             final long packedPosition = ExpandableAdapterHelper.getPackedPositionForGroup(groupPosition);
             final int flatPosition = mPositionTranslator.getFlatPosition(packedPosition);
 
-            notifyItemChanged(flatPosition);
+            notifyItemChanged(flatPosition, payload);
         }
 
         // raise onGroupExpand() event
         if (mOnGroupExpandListener != null) {
-            mOnGroupExpandListener.onGroupExpand(groupPosition, fromUser);
+            mOnGroupExpandListener.onGroupExpand(groupPosition, fromUser, payload);
         }
 
         return true;
@@ -923,7 +935,7 @@ class ExpandableRecyclerViewWrapperAdapter
             notifyItemInserted(flatPosition);
 
             // raise onGroupExpand() event
-            raiseOnGroupExpandedSequentially(groupPosition, 1, false);
+            raiseOnGroupExpandedSequentially(groupPosition, 1, false, null);
         }
     }
 
@@ -935,7 +947,7 @@ class ExpandableRecyclerViewWrapperAdapter
 
             notifyItemRangeInserted(flatPosition, insertedCount);
 
-            raiseOnGroupExpandedSequentially(groupPositionStart, count, false);
+            raiseOnGroupExpandedSequentially(groupPositionStart, count, false, null);
         }
     }
 
@@ -977,10 +989,10 @@ class ExpandableRecyclerViewWrapperAdapter
         }
     }
 
-    private void raiseOnGroupExpandedSequentially(int groupPositionStart, int count, boolean fromUser) {
+    private void raiseOnGroupExpandedSequentially(int groupPositionStart, int count, boolean fromUser, Object payload) {
         if (mOnGroupExpandListener != null) {
             for (int i = 0; i < count; i++) {
-                mOnGroupExpandListener.onGroupExpand(groupPositionStart + i, fromUser);
+                mOnGroupExpandListener.onGroupExpand(groupPositionStart + i, fromUser, payload);
             }
         }
     }